x4_app/pyproject.toml
[project]
name = "x4_app"
version = "1.0.0"
description = "TODO"
authors = [{name = "Ryan D. Lewis", email="ryan.lewis@anl.gov"}]
dependencies = [
    "ad_sdl.wei>=0.6.2",
]
requires-python = ">=3.8.1"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/AD-SDL/x4_workcell"
